moved client classes into separate files, created DataClient and associated buffers
