+++ RemoteApplication/Network/control_client.py
@@ -1,0 +1,94 @@
+import socket
+import multiprocessing as mp
+import threading
+import asyncore
+import sys
+import control_signals_pb2
+
+class ControlClient(asyncore.dispatcher):
+    def __init__(self, host, port, outgoing_queue, incoming_queue):
+        asyncore.dispatcher.__init__(self)
+
+        # initialize TCP socket
+        self.create_socket(socket.AF_INET, socket.SOCK_STREAM)
+
+        # holds messages that have not been sent yet
+        self.outgoing_queue = outgoing_queue
+
+        # holds serialized messages that have been received but not processed
+        self.incoming_queue = incoming_queue
+
+        # holds messages that have been sent over the
+        # control_socket but have not yet been ACKed
+        self.sent_dict = {}
+
+        self.host = host
+        self.port = port
+        self.connected = False
+
+    def connect_control_port(self):
+        print 'attempting connection'
+        self.connect((self.host, self.port))
+        self.connected = True
+
+    def close_control_port(self):
+        print 'ControlClient: close_control_port()'
+        self.connected = False
+        self.close()
+
+    def handle_connect(self):
+        print 'handle_connect() entered'
+
+    def handle_close(self):
+        self.connected = False
+        self.close()
+
+    def handle_read(self):
+        # read 16 bit length header
+        length = self.recv(2)
+        print 'received length header: %s' % length
+
+        # read the message content
+        msg = self.recv(length)
+        print 'received message: %s' % msg
+
+        # TODO: put onto incoming_queue and have another thread handle the parsing
+
+        # construct a container protobuf and parse into from serialized message
+        ackWrapper = control_signals_pb2.RequestWrapper()
+        ackWrapper.ParseFromString(msg)
+
+        seq = ackWrapper.seq
+
+        if seq in self.sent_dict.keys():
+            acked_request = self.sent_dict.pop(seq)
+            print 'ACKed request popped %s' % acked_request
+            # TODO: process ACK and notify necessary parties
+
+    def readable(self):
+        return True
+
+    def writable(self):
+        # we want to write whenever there are messages to be sent
+        is_writable = not self.outgoing_queue.empty()
+        return is_writable
+
+    def handle_write(self):
+        # grab request to be sent from the queue
+        serialized_req_wrap = self.outgoing_queue.get_nowait()
+        print 'handle_write() retrieved msg from outgoing queue'
+
+        # parse the request for storage in sent_dict
+        request_wrapper = control_signals_pb2.RequestWrapper()
+        request_wrapper.ParseFromString(serialized_req_wrap)
+
+        print 'sending message length over control socket'
+        self.send(str(sys.getsizeof(serialized_req_wrap)))
+        # TODO: sent as a uint16
+
+        print 'sending Request message over control socket'
+        sent = self.send(serialized_req_wrap)
+        print 'sent message bytes: %d' % sent
+
+        print 'adding request %d to sent_dict' % request_wrapper.sequence
+        self.sent_dict[request_wrapper.sequence] = request_wrapper
